+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 Change Log
 ==========
 
+2020-09-29
+----------
+
+ * Re-namespace to `Somnambulist\Components`
+
 2020-08-26
 ----------
 
-Require PHP 7.4 as minimum version
+ * Require PHP 7.4 as minimum version
 
 2019-08-15
 ----------
 
-Initial commit
+ * Initial commit